perf(fock): Faster `FockState.norm`
A custom `norm` method has been written for `FockState`, which is faster with
`TensorflowCalculator`, since it uses less tensorflow operations. Moreover, the
typing in the signature of `norm` is fixed along the way.

+++ piquasso/_backends/fock/state.py
@@ -38,7 +38,7 @@ class BaseFockState(State, abc.ABC):
         return self._d
 
     @property
-    def norm(self) -> int:
+    def norm(self) -> float:
         return self._calculator.np.sum(self.fock_probabilities)
 
     def _as_code(self) -> str:
